Tweaks to shooter, climber, drive flip, and new shooter cancel
Results from testing the robot today. The limelight RPM calculation is functional, and the shooter is tuned with voltage compensation for consistency. The feeder logic was ironed out and tested, and the control for flipping the joysticks, as well as cancelling shooting were implemented. Auto climb was also removed, and haptic feedback implemented in this code.