Shooter, limelight, and auto work
(Jacob, Sushrit, Eric) Worked on retuning the shooter, as well as the angle PID system. Also touched a bit on the feeder speeds pending the changes to the reduction on the feeder wheels. The shooter feeder integration runs on a RPM error system, rather than a constant feed system as before. Finally, fixed lack of auto conversion in setpoint.