towerSimpleRamseteCommand Finished
Implemented the RamseteCommand for towerSimple for testing tomorrow, if it works then we know we can use this approach for the other autonomous trajectories. One issues in this are that when im requiring the drive subsystem in the ramsete command its giving an error. I also need to add the line of code in RobotContainer under the TowerSimple case to return the autonomous command from TowerSimple.java